+++ enclave/src/inference.rs
@@ -3,23 +3,23 @@ use crate::assembly::CandleModel;
 
 /// Trait for inference execution
 pub trait InferenceEngine {
-    /// Execute inference on a model with input data
-    fn execute(&self, model: &CandleModel, input: &[f32]) -> Result<Vec<f32>>;
+    /// Execute inference on a model with input data (as bytes for flexibility)
+    fn execute(&self, model: &CandleModel, input: &[u8]) -> Result<Vec<f32>>;
     
     /// Validate input data shape and format
-    fn validate_input(&self, model: &CandleModel, input: &[f32]) -> Result<()>;
+    fn validate_input(&self, model: &CandleModel, input: &[u8]) -> Result<()>;
 }
 
 /// Default inference engine implementation
 pub struct DefaultInferenceEngine;
 
 impl InferenceEngine for DefaultInferenceEngine {
-    fn execute(&self, _model: &CandleModel, _input: &[f32]) -> Result<Vec<f32>> {
+    fn execute(&self, _model: &CandleModel, _input: &[u8]) -> Result<Vec<f32>> {
         // Placeholder implementation - will be implemented in later tasks
         Err(EnclaveError::Enclave(EphemeralError::InferenceError("Not yet implemented".to_string())))
     }
     
-    fn validate_input(&self, _model: &CandleModel, _input: &[f32]) -> Result<()> {
+    fn validate_input(&self, _model: &CandleModel, _input: &[u8]) -> Result<()> {
         // Placeholder implementation - will be implemented in later tasks
         Ok(())
     }
